backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
numpy
pandas
networkx
scikit-learn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
python-dotenv
fastapi-mail
ollama
pypdf